pyproject.toml
[tool.poetry]
name = "jwt"
version = "0.1.0"
description = "JWT service"
authors = ["Rubnikovich Sergey <https://github.com/Sergey-x>"]

[tool.poetry.dependencies]
python = "3.8.10"
fastapi = "0.75.1"
uvicorn = { extras = ["standard"], version = "0.17.6" }
PyJWT = "2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
